+++ src/proto/proto.py
@@ -1,27 +1,23 @@
 import logging
-import importlib
-from zensols.nlp import AppConfig
 
 logger = logging.getLogger(__name__)
 
 
-def create_config():
-    return AppConfig('resources/nlparse.conf')
-
-
-def tmp():
-    import zensols.nlp.app
-    importlib.reload(zensols.nlp.app)
-    logging.getLogger('zensols.actioncli').setLevel(logging.INFO)
-    logging.getLogger('zensols.nlp.app').setLevel(logging.DEBUG)
-    app = zensols.nlp.app.MainApplication(create_config())
-    app.tmp()
+def test():
+    from zensols.actioncli import ClassImporter
+    import sys
+    sys.path.append('./test/python')
+    inst = ClassImporter('test_nlparse.TestParse').instance()
+    #logging.getLogger('zensols.nlp').setLevel(level=logging.DEBUG)
+    inst.setUp()
+    inst.test_disable()
+    inst.tearDown()
 
 
 def main():
     logging.basicConfig(level=logging.WARNING)
     run = 1
-    {1: tmp,
+    {1: test,
      }[run]()